django/requirements.txt
Django
djangorestframework
markdown
django-filter
coverage
Django-cors-headers